Migration trait unified C*, Zk transformation
