+++ src/main/scala/net/elodina/mesos/dse/CassandraStorage.scala
@@ -24,7 +24,7 @@ import scala.collection.JavaConverters._
 import scala.collection.mutable.ListBuffer
 import scala.util.Try
 
-class CassandraStorage(port: Int, contactPoints: Seq[String], keyspace: String, stateTable: String) extends Storage {
+class CassandraStorage(port: Int, contactPoints: Seq[String], keyspace: String, stateTable: String, versionTable: String = "version") extends Storage {
 
   import CassandraStorage._
 
@@ -35,7 +35,8 @@ class CassandraStorage(port: Int, contactPoints: Seq[String], keyspace: String, 
       .withPort(port).addContactPoints(contactPoints: _*).build().connect(keyspace)
 
   private val SelectPs = session.prepare(CassandraStorage.selectQuery(stateTable))
-  private val InsertionPs = session.prepare(CassandraStorage.insertionQuery(stateTable))
+  // have to make InsertionPs lazy because it uses fields that could not exist at the moment (e.g. added after migration)
+  private lazy val InsertionPs = session.prepare(CassandraStorage.insertionQuery(stateTable))
   private val DeletionPs = session.prepare(CassandraStorage.deleteQuery(stateTable))
 
   private def stringOrNull[T](value: T): String = {
@@ -256,6 +257,8 @@ class CassandraStorage(port: Int, contactPoints: Seq[String], keyspace: String, 
   }
 
   override def load(): Boolean = {
+    migrate
+
     val boundStatement = SelectPs.bind().setString(Namespace, Config.namespace).setConsistencyLevel(ConsistencyLevel.ONE)
 
     val rows = session.execute(boundStatement).all().asScala
@@ -300,6 +303,28 @@ class CassandraStorage(port: Int, contactPoints: Seq[String], keyspace: String, 
       if (session != null)
         session.close()
     }
+  }
+
+  private[dse] def migrate: Unit = {
+    val verTable = session.getCluster.getMetadata.getKeyspace(keyspace).getTable(versionTable)
+    if (verTable == null) {
+      session.execute(s"create table $keyspace.$versionTable (latest text PRIMARY KEY)")
+      session.execute(s"insert into $keyspace.$versionTable (latest) values ('0.2.1.2')")
+    }
+
+    val schemaVersion: Version = {
+      val result = session.execute(s"select latest from $keyspace.$versionTable")
+      new Version(result.one().getString("latest"))
+    }
+
+    def updateVersion(v: Version): Unit = {
+      session.execute(s"truncate $keyspace.$versionTable")
+      session.execute(s"insert into $keyspace.$versionTable (latest) values ('$v')")
+    }
+
+    Migrator.migrate(schemaVersion, Scheduler.version, Migrator.migrations, updateVersion, m => m.migrateCassandra(session))
+
+    updateVersion(Scheduler.version)
   }
 }
 
@@ -411,92 +436,4 @@ object CassandraStorage{
 
   def selectQuery(table: String) =
     s"SELECT * FROM $table WHERE $Namespace = ${`:`(Namespace)}"
-}
-
-class CassandraStorageMigrator(port: Int, contactPoints: Seq[String], keyspace: String, stateTable: String) {
-  val versionTableName = "version"
-
-  def migrate: Unit = {
-    val session = Cluster.builder().withPort(port).addContactPoints(contactPoints: _*).build().connect(keyspace)
-    try {
-      val versionTable = session.getCluster.getMetadata.getKeyspace(keyspace).getTable(versionTableName)
-      if (versionTable == null) {
-        session.execute(s"create table $keyspace.$versionTableName (latest text PRIMARY KEY)")
-        session.execute(s"insert into $keyspace.$versionTableName (latest) values ('0.2.1.2')")
-      }
-
-      val schemaVersion: Version = {
-        val result = session.execute(s"select latest from $keyspace.$versionTableName")
-        new Version(result.one().getString("latest"))
-      }
-
-      val migrations = Seq(
-        new CassandraMigration0_2_1_3(session, keyspace, stateTable)
-      )
-
-      def updateVersion(v: Version): Unit = {
-        session.execute(s"truncate $keyspace.$versionTableName")
-        session.execute(s"insert into $keyspace.$versionTableName (latest) values ('$v')")
-      }
-
-      Migrator.migrate(schemaVersion, Scheduler.version, migrations, updateVersion)
-
-      updateVersion(Scheduler.version)
-    } finally {
-      session.close()
-    }
-  }
-
-}
-
-class CassandraMigration0_2_1_3(session: Session, keyspace: String, stateTable: String) extends Migration {
-  import CassandraStorage._
-
-  override val version: Version = new Version("0.2.1.3")
-
-  val alters = Seq(
-    s"alter table $keyspace.$stateTable add cluster_jmx_remote boolean",
-    s"alter table $keyspace.$stateTable add cluster_jmx_user text",
-    s"alter table $keyspace.$stateTable add cluster_jmx_password text",
-
-    s"alter table $keyspace.$stateTable add node_failover_delay text",
-    s"alter table $keyspace.$stateTable add node_failover_max_delay text",
-    s"alter table $keyspace.$stateTable add node_failover_max_tries int",
-    s"alter table $keyspace.$stateTable add node_failover_failures int",
-    s"alter table $keyspace.$stateTable add node_failover_failure_time timestamp"
-  )
-
-  override def apply(): Unit = {
-    alters.foreach(session.execute)
-
-    val updatePs = session.prepare(s"""
-            UPDATE $keyspace.$stateTable
-            SET
-            cluster_jmx_remote = false,
-            node_failover_delay = '3m',
-            node_failover_max_delay = '30m',
-            node_failover_failures = 0
-            WHERE
-            namespace = :namespace AND
-              framework_id = :framework_id AND
-              cluster_id = :cluster_id AND
-              node_id = :node_id
-          """)
-
-    val selectPs = session.prepare(s"select namespace, framework_id, cluster_id, node_id, nr_of_nodes from $keyspace.$stateTable")
-
-    val batch = new BatchStatement()
-    val rows = session.execute(selectPs.bind()).all().asScala
-    for (row <- rows) {
-      val update = updatePs.bind()
-        .setString(Namespace, row.getString(Namespace))
-        .setString(FrameworkId, row.getString(FrameworkId))
-        .setString(ClusterId, row.getString(ClusterId))
-        .setString(NodeId, row.getString(NodeId))
-
-      batch add update
-    }
-
-    session execute batch
-  }
 }
