+++ src/main/scala/net/elodina/mesos/dse/Migration.scala
@@ -18,27 +18,21 @@
 
 package net.elodina.mesos.dse
 
-import java.io.File
-import org.I0Itec.zkclient.ZkClient
-import org.I0Itec.zkclient.exception.ZkNodeExistsException
-import org.I0Itec.zkclient.serialize.BytesPushThroughSerializer
+import scala.collection.JavaConverters._
 import net.elodina.mesos.dse.Util.Version
-import scala.util.parsing.json.JSONObject
+import com.datastax.driver.core._
 
 trait Migration {
   val version: Version
-  def apply(): Unit
-}
-
-trait Migrator {
-  def migrate: Unit
+  def migrateJson(json: Map[String, Any]): Map[String, Any]
+  def migrateCassandra(session: Session): Unit
 }
 
 object Migrator {
   import org.apache.log4j.Logger
   private val logger = Logger.getLogger(Migrator.getClass)
 
-  def migrate(from: Version, to: Version, migrations: Seq[Migration], updateVersion: Version => Unit): Unit = {
+  def migrate(from: Version, to: Version, migrations: Seq[Migration], updateVersion: Version => Unit, apply: Migration => Unit): Unit = {
     if (from == to) {
       logger.info("storage schema is up to date")
       return
@@ -52,52 +46,21 @@ object Migrator {
 
     for (m <- applicable) {
       logger.info("applying migration for version " + m.version)
-      m()
+      apply(m)
       updateVersion(m.version)
     }
     logger.info(s"migration from $from to $to completed")
   }
+
+  val migrations = Seq[Migration](
+    new Migration0_2_1_3(Config.cassandraKeyspace, Config.cassandraTable)
+  )
 }
 
-class FileStorageMigrator(file: File) extends Migrator {
-  def migrate = {
-    if (file.exists()) {
-      var schemaVersion: Version = null
-      withJson { json =>
-        val versioned = if (!json.contains("version")) json.updated("version", "0.2.1.2") else json
-        schemaVersion = new Version(versioned("version").asInstanceOf[String])
-        versioned
-      }
+class Migration0_2_1_3(keyspace: String, stateTable: String) extends Migration {
+  override val version: Version = new Version("0.2.1.3")
 
-      val migrations = Seq(
-        mkMigration("0.2.1.3")(JSONMigrations.v0_2_1_3)
-      )
-
-      def updateVersion(v: Version): Unit = withJson { json => json.updated("version", v.toString) }
-
-      Migrator.migrate(schemaVersion, Scheduler.version, migrations, updateVersion)
-
-      updateVersion(Scheduler.version)
-    }
-  }
-
-  def withJson(f: Map[String, Any] => Map[String, Any]): Unit = {
-    if (file.exists()) {
-      val content = Util.IO.readFile(file)
-      val json = Util.parseJsonAsMap(content)
-      Util.IO.writeFile(file, new JSONObject(f(json)).toString(Util.jsonFormatter))
-    }
-  }
-
-  def mkMigration(v: String)(f: Map[String, Any] => Map[String, Any]): Migration = new Migration {
-    override val version: Version = new Version(v)
-
-    override def apply(): Unit = withJson(f)
-  }
-}
-
-object JSONMigrations {
-  def v0_2_1_3(json: Map[String, Any]): Map[String, Any] = {
+  override def migrateJson(json: Map[String, Any]): Map[String, Any] = {
     // update cluster, add jmxRemote false
     val json1 = if (json.contains("clusters")) {
       val clusters = json("clusters").asInstanceOf[List[Map[String, Object]]]
@@ -112,68 +75,54 @@ object JSONMigrations {
 
     json2
   }
-}
 
-trait ZkMigratorHelpers {
-  val path: String
+  val alters = Seq(
+    s"alter table $keyspace.$stateTable add cluster_jmx_remote boolean",
+    s"alter table $keyspace.$stateTable add cluster_jmx_user text",
+    s"alter table $keyspace.$stateTable add cluster_jmx_password text",
 
-  def withJson(client: ZkClient)(f: Map[String, Any] => Map[String, Any]) {
-    val json = {
-      val bytes: Array[Byte] = client.readData(path, true).asInstanceOf[Array[Byte]]
-      if (bytes != null) Util.parseJsonAsMap(new String(bytes, "utf-8"))
-      else Map.empty[String, Any]
+    s"alter table $keyspace.$stateTable add node_failover_delay text",
+    s"alter table $keyspace.$stateTable add node_failover_max_delay text",
+    s"alter table $keyspace.$stateTable add node_failover_max_tries int",
+    s"alter table $keyspace.$stateTable add node_failover_failures int",
+    s"alter table $keyspace.$stateTable add node_failover_failure_time timestamp"
+  )
+
+  override def migrateCassandra(session: Session): Unit = {
+    alters.foreach(session.execute)
+
+    val updatePs = session.prepare(s"""
+          UPDATE $keyspace.$stateTable
+          SET
+          cluster_jmx_remote = false,
+          node_failover_delay = '3m',
+          node_failover_max_delay = '30m',
+          node_failover_failures = 0
+          WHERE
+          namespace = :namespace AND
+            framework_id = :framework_id AND
+            cluster_id = :cluster_id AND
+            node_id = :node_id
+        """)
+
+    val selectPs = session.prepare(s"select namespace, framework_id, cluster_id, node_id, nr_of_nodes from $keyspace.$stateTable")
+
+    import CassandraStorage._
+
+    val batch = new BatchStatement()
+    batch.setConsistencyLevel(ConsistencyLevel.ONE)
+    val rows = session.execute(selectPs.bind()).all().asScala
+    for (row <- rows) {
+      val update = updatePs.bind()
+        .setString(Namespace, row.getString(Namespace))
+        .setString(FrameworkId, row.getString(FrameworkId))
+        .setString(ClusterId, row.getString(ClusterId))
+        .setString(NodeId, row.getString(NodeId))
+
+      batch add update
     }
 
-    val json1 = f(json)
-
-    val encoded = new JSONObject(json1).toString(Util.jsonFormatter).getBytes("utf-8")
-    try {
-      client.createPersistent(path, encoded)
-    } catch {
-      case e: ZkNodeExistsException => client.writeData(path, encoded)
-    }
-  }
-}
-
-class ZkStorageMigrator(zk: String) extends Migrator with ZkMigratorHelpers {
-  val (zkConnect, path) = zk.span(_ != '/')
-
-  override def migrate: Unit = {
-    withClient { client =>
-      if (path != "") client.createPersistent(path, true)
-
-      var schemaVersion: Version = null
-      withJson(client) { json =>
-        val versioned = if (!json.contains("version")) json.updated("version", "0.2.1.2") else json
-        schemaVersion = new Version(versioned("version").asInstanceOf[String])
-        versioned
-      }
-
-      val migrations = Seq(
-        new ZkMigraiton0_2_1_3(client, path)
-      )
-
-      def updateVersion(v: Version) = withJson(client) { json => json.updated("version", v.toString) }
-
-      Migrator.migrate(schemaVersion, Scheduler.version, migrations, updateVersion)
-
-      updateVersion(Scheduler.version)
-    }
-  }
-
-  private[dse] def withClient[T](f: ZkClient => T): T = {
-    val c = zkClient
-    try { f(c) } finally { c.close() }
-  }
-
-  private[dse] def zkClient: ZkClient = new ZkClient(zkConnect, 30000, 30000, new BytesPushThroughSerializer)
-}
-
-class ZkMigraiton0_2_1_3(client: ZkClient, override val path: String) extends Migration with ZkMigratorHelpers {
-  override val version: Version = new Version("0.2.1.3")
-
-  override def apply(): Unit = {
-    withJson(client)(JSONMigrations.v0_2_1_3)
+    session execute batch
   }
 }
