+++ src/main/scala/net/elodina/mesos/dse/Storage.scala
@@ -22,6 +22,8 @@ import java.io.File
 import org.I0Itec.zkclient.ZkClient
 import org.I0Itec.zkclient.exception.ZkNodeExistsException
 import org.I0Itec.zkclient.serialize.BytesPushThroughSerializer
+import net.elodina.mesos.dse.Util.Version
+import scala.util.parsing.json.JSONObject
 
 trait Storage {
   def save()
@@ -45,10 +47,42 @@ case class FileStorage(file: File) extends Storage {
 
   override def load(): Boolean = {
     if (file.exists()) {
+      migrate
       val json = Util.IO.readFile(file)
       Nodes.fromJson(Util.parseJsonAsMap(json))
       true
     } else false
+  }
+
+  private[dse] def migrate: Unit = {
+    if (file.exists()) {
+      var schemaVersion: Version = null
+      withJson { json =>
+        val versioned = if (!json.contains("version")) json.updated("version", "0.2.1.2") else json
+        schemaVersion = new Version(versioned("version").asInstanceOf[String])
+        versioned
+      }
+
+      def updateVersion(v: Version): Unit = withJson { json => json.updated("version", v.toString) }
+
+      var json: Map[String, Any] = readJson
+      Migrator.migrate(schemaVersion, Scheduler.version, Migrator.migrations, v => (), m => json = m.migrateJson(json))
+      writeJson(json)
+
+      updateVersion(Scheduler.version)
+    }
+  }
+
+  def readJson: Map[String, Any] =
+    if (file.exists()) Util.parseJsonAsMap(Util.IO.readFile(file))
+    else Map.empty[String, Any]
+
+  def writeJson(json: Map[String, Any]): Unit = {
+    Util.IO.writeFile(file, new JSONObject(json).toString(Util.jsonFormatter))
+  }
+
+  def withJson(f: Map[String, Any] => Map[String, Any]): Unit = {
+    if (file.exists()) Util.IO.writeFile(file, new JSONObject(f(readJson)).toString(Util.jsonFormatter))
   }
 }
 
@@ -56,47 +90,68 @@ case class ZkStorage[T](zk: String) extends Storage {
   val (zkConnect, path) = zk.span(_ != '/')
   createChrootIfRequired()
 
-  private def createChrootIfRequired() {
-    if (path != "") {
-      val client = zkClient
-      try {
-        client.createPersistent(path, true)
-      }
-      finally {
-        client.close()
-      }
-    }
-  }
-
-  private def zkClient: ZkClient = new ZkClient(zkConnect, 30000, 30000, new BytesPushThroughSerializer)
-
   override def save() {
-    val json = Nodes.toJson
-    val client = zkClient
-    val encoded = json.toString().getBytes("utf-8")
-    try {
-      client.createPersistent(path, encoded)
-    } catch {
-      case e: ZkNodeExistsException => client.writeData(path, encoded)
-    }
-    finally {
-      client.close()
-    }
+    withClient { client => write(client, Nodes.toJson.toString().getBytes("utf-8")) }
   }
 
   override def load(): Boolean = {
-    val client = zkClient
-    try {
-      val bytes: Array[Byte] = client.readData(path, true).asInstanceOf[Array[Byte]]
-      if (bytes != null) {
-        val map = Util.parseJsonAsMap(new String(bytes, "utf-8"))
-        Nodes.fromJson(map)
+    migrate
+
+    withClient { client =>
+      val json = readJson(client)
+      if (json.nonEmpty) {
+        Nodes.fromJson(json)
         true
       } else false
-
-    } finally {
-      client.close()
     }
   }
+
+  private[dse] def migrate: Unit = {
+    withClient { client =>
+      var schemaVersion: Version = null
+      withJson(client) { json =>
+        val versioned = if (!json.contains("version")) json.updated("version", "0.2.1.2") else json
+        schemaVersion = new Version(versioned("version").asInstanceOf[String])
+        versioned
+      }
+
+      def updateVersion(v: Version) = withJson(client) { json => json.updated("version", v.toString) }
+
+      var json = readJson(client)
+      Migrator.migrate(schemaVersion, Scheduler.version, Migrator.migrations, v => (), m => json = m.migrateJson(json))
+      writeJson(client, json)
+
+      updateVersion(Scheduler.version)
+    }
+  }
+
+  private def createChrootIfRequired(): Unit = if (path != "") withClient(_.createPersistent(path, true))
+
+  private[dse] def zkClient: ZkClient = new ZkClient(zkConnect, 30000, 30000, new BytesPushThroughSerializer)
+
+  private[dse] def withClient[T](f: ZkClient => T): T = {
+    val c = zkClient
+    try { f(c) } finally { c.close() }
+  }
+
+  private def write(client: ZkClient, bytes: Array[Byte]) {
+    try {
+      client.createPersistent(path, bytes)
+    } catch {
+      case e: ZkNodeExistsException => client.writeData(path, bytes)
+    }
+  }
+
+  private def readJson(client: ZkClient): Map[String, Any] = {
+    val bytes: Array[Byte] = client.readData(path, true).asInstanceOf[Array[Byte]]
+    if (bytes != null) Util.parseJsonAsMap(new String(bytes, "utf-8"))
+    else Map.empty[String, Any]
+  }
+
+  private def writeJson(client: ZkClient, json: Map[String, Any]): Unit = {
+    write(client, new JSONObject(json).toString(Util.jsonFormatter).getBytes("utf-8"))
+  }
+
+  private def withJson(client: ZkClient)(f: Map[String, Any] => Map[String, Any]): Unit = writeJson(client, f(readJson(client)))
 }
